feat: Update pre-build-buildpack to install latest version of chromium
This commit updates the pre-build-buildpack script to install the latest version of chromium using the Google Chrome browser snapshots. The script now fetches the latest revision number and downloads the corresponding zip file from the snapshots repository. The downloaded zip file is then unzipped and symlinked as the latest version of chromium. This change ensures that the Elixir project has the most up-to-date version of chromium installed.